Add tests for uploadCard-component
